feat(interpreter): allow instructions to be implemented without a Host generic
Reduces monomorphization bloat by reusing the same `fn` item for all `H: Host`s.
This is especially a problem for the functions which were generic over both `H`
and `SPEC`, as often hundreds of copies are created of the same function.

Should reduce compile times in downstream users of Revm.